love you simon
everything will be okay! the game will get finished and it will be something we are all proud of, and whether we finish tonight or next year it is totally fine, me and nate love you and if you are overwhelmed we dont have to finish, we would rather you be okay then be stressed by this game.